test : 음식점 추가 페이지 테스트 작성
